requirements.txt
PySide6_Essentials;
requests;
urllib3;
setproctitle;
psutil;
yt-dlp;
